Rename files and move code to better named files. DO NOT SQUASH COMMIT.
